Introduce provided values in v2
In v1 it was possible to take into account another value from the hash
when validating some particular value. V2 was lacking this behavior.

This commit introduces that into v2. In case of nested converters we
have to take care of provided key from the validtion context. The way it
works now is that it merges the provided values from context and from
the current converter itself before going to clean field routine.